requirements.txt
django
south
django-debug-toolbar
django-annoying
